Add comments and Change Naming Convention
